nppez: Tidy NPPES NPI Registry

Provides a tidy interface to the National Plan and Provider Enumeration System (NPPES) National Provider Identifier (NPI) Registry, a database of all health care providers who have been assigned an NPI. Includes functions to download and parse the NPI Registry data, perform updates according to the NPPES data release schedule and prepare it for storage.

See also

Author

Maintainer: Andrew Bruce andrewallenbruce@gmail.com [copyright holder]